clean up python
